standalone rpc requesert
