+++ src/helpers/evm/evm.ts
@@ -1,25 +1,30 @@
-import WalletConnect from "@walletconnect/client";
 import { InjectedProvider, RequestArguments } from "types/evm";
 import { ProviderId } from "types/lists";
 import { Dwindow } from "types/window";
 import web3Auth from "contexts/WalletContext/useWeb3Auth/web3AuthSetup";
-import { WC_BRIDGE } from "constants/urls";
+import { _session, account } from "helpers/wallet-connect";
+import { POLYGON_RPC } from "constants/urls";
 
-export const connector = new WalletConnect({
-  bridge: WC_BRIDGE,
-  storageId: "wc_evm",
-  qrcodeModalOptions: { mobileLinks: ["metamask"], desktopLinks: [] },
-});
-
-const wcProvider: Partial<InjectedProvider> = {
-  async request<T>({ method, params }: RequestArguments): Promise<T> {
-    return connector.sendCustomRequest({ method, params: params as any });
-  },
-};
-
-export function getProvider(
+export async function wcProvider(): Promise<Partial<InjectedProvider>> {
+  //FUTURE: pass peer name in wcProvider call
+  const { session, client } = await _session("MetaMask Wallet");
+  const { chainId } = account(session!.namespaces.eip155);
+  return {
+    async request<T>({ method, params }: RequestArguments): Promise<T> {
+      return client.request<T>({
+        topic: session!.topic,
+        chainId: `eip155:${chainId}`,
+        request: {
+          method,
+          params,
+        },
+      });
+    },
+  };
+}
+export async function getProvider(
   providerId: ProviderId
-): InjectedProvider | undefined {
+): Promise<InjectedProvider | undefined> {
   const dwindow = window as Dwindow;
   switch (providerId) {
     case "binance-wallet":
@@ -30,10 +35,31 @@ export function getProvider(
       return web3Auth.provider as InjectedProvider;
     /** only used in sendTx */
     case "evm-wc":
-      return wcProvider as InjectedProvider;
+      return wcProvider() as Promise<InjectedProvider>;
     case "xdefi-evm":
       return dwindow.xfi?.ethereum as InjectedProvider;
     default:
       return undefined;
   }
+}
+
+type Result = { result: string } | { error: { code: number; message: string } };
+export async function request({ method, params }: RequestArguments) {
+  const result = await fetch(POLYGON_RPC, {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify({
+      jsonrpc: "2.0",
+      id: 1,
+      method,
+      params,
+    }),
+  }).then<Result>((res) => {
+    if (!res.ok) throw new Error(`Failed request ${method}`);
+    return res.json();
+  });
+
+  if ("error" in result) throw new Error(result.error.message);
+
+  return result.result;
 }
