+++ src/helpers/tx/estimateTx/estimateEVMfee.ts
@@ -4,26 +4,24 @@ import { Estimate } from "types/tx";
 import { WalletState } from "contexts/WalletContext";
 import { EIPMethods } from "constants/evm";
 import { condense } from "../../decimal";
-import { getProvider } from "../../evm";
+import { request } from "../../evm";
 
 export async function estimateEVMFee(
   wallet: WalletState,
   tx: SimulTx,
   log?: LogProcessor
 ): Promise<Estimate> {
-  const provider = getProvider(wallet.providerId)!;
-
   //TODO: convert request to fetch call so can also be used in general contract queries
   const [nonce, gas, gasPrice] = await Promise.all([
-    provider.request<string>({
+    request({
       method: EIPMethods.eth_getTransactionCount,
       params: [wallet.address, "latest"],
     }),
-    provider.request<string>({
+    request({
       method: EIPMethods.eth_estimateGas,
       params: [tx],
     }),
-    provider.request<string>({
+    request({
       method: EIPMethods.eth_gasPrice,
     }),
   ]);
